feat: added footer on all pages
