+++ src/assets/scss/footer.scss
@@ -6,33 +6,9 @@ footer {
 	width: 100vw;
 	max-width: 100vw;
 	min-width: 100vw;
+	margin-bottom: -2em;
 	@include themify() {
 		background: themed("footerBackgroundColor");
-	}
-}
-
-.socials {
-	display: grid;
-	grid-template-columns: auto auto auto;
-	.twitter,
-	.discord {
-		flex-grow: 2;
-	}
-	.twitter,
-	.discord,
-	.github {
-		padding: 1em 1em;
-		display: grid;
-		place-items: center;
-		color: inherit;
-		text-decoration: none;
-		a {
-			text-align: center;
-			color: inherit;
-		}
-		svg {
-			font-size: 1.5em;
-		}
 	}
 }
