+++ src/views/Home.vue
@@ -40,7 +40,27 @@
 				</div>
 			</div>
 		</main>
-		<Footer />
+		<div class="socials">
+			<div class="discord">
+				<a href="https://discord.com/invite/k7b27z77" target="popup">
+					<DiscordLogo />
+					<p>Join 7TV on Discord</p>
+					<i>891 members online now</i>
+				</a>
+			</div>
+			<div class="twitter">
+				<a href="https://twitter.com/Official_7TV" target="_BLANK">
+					<font-awesome-icon :icon="['fab', 'twitter']" />
+					<p>Follow us on Twitter!</p>
+				</a>
+			</div>
+			<div class="github">
+				<a href="https://github.com/SevenTV" target="_BLANK">
+					<font-awesome-icon :icon="['fab', 'github']" />
+					<p>Contribute</p>
+				</a>
+			</div>
+		</div>
 	</div>
 </template>
 
@@ -51,14 +71,15 @@ import { defineComponent } from "vue";
 import Logo from "@/components/base/Logo.vue";
 import ChatterinoLogo from "@/components/base/ChatterinoLogo.vue";
 import FirefoxLogo from "@/components/base/FirefoxLogo.vue";
-import Footer from "@/components/Footer.vue";
+
+import DiscordLogo from "@/components/base/DiscordLogo.vue";
 
 export default defineComponent({
 	components: {
 		Logo,
+		DiscordLogo,
 		ChatterinoLogo,
 		FirefoxLogo,
-		Footer,
 	},
 	setup() {
 		useHead({
